input sanitation
we now have input sanitization! please play around and try to break it, i think it should work. also played around with gradient -> i might try to make a new function which creates a list for the palette based on the answers, and then passes the list around as an argument (instead of calling the randomColor function a million times). this would probably be more efficient, but would change how a lot of our functions look.